+++ analytics/base/src/main/java/edu/jhu/hlt/concrete/analytics/base/Analytic.java
@@ -6,18 +6,18 @@ package edu.jhu.hlt.concrete.analytics.base;
 
 import edu.jhu.hlt.concrete.Communication;
 import edu.jhu.hlt.concrete.metadata.tools.SafeTooledAnnotationMetadata;
+import edu.jhu.hlt.concrete.miscommunication.WrappedCommunication;
 
 /**
  * Top level interface for Concrete analytics. Loosely defined as
  * taking in a {@link Communication} and outputting a {@link Communication}
  * with some additions.
  */
-public interface Analytic extends SafeTooledAnnotationMetadata {
+public interface Analytic<T extends WrappedCommunication> extends SafeTooledAnnotationMetadata {
   /**
    * @param c a {@link Communication} to annotate. It is up to the analytic to determine
    * if this communication is valid for annotation.
-   * @return a {@link Communication} object with the analytic's annotations
    * @throws AnalyticException on analytic error
    */
-  public Communication annotate(Communication c) throws AnalyticException;
+  public T annotate(Communication c) throws AnalyticException;
 }